Gif code dump
The code is not pretty.  I became increasingly rushed as the deadline
approached.  But if anyone is curious how to make animations like these,
everything they need *should* be here, once all the necessary libraries
are installed.

Oh, and gppois is no longer maintained (sorry!), so you have to
install_github() it (using devtools).